rename some files and classes mentionning twitter
