+++ infra/ansible/scripts/forget-secrets.sh
