+++ infra/ansible/scripts/get-vm-secrets.sh
